patient matching module: performance tuning by removing code which will send the report to the web page, using ReaderProvider class to get the scratch table reader, AJAX-ing the report page, remove redundant serialization when a record is already serialized before.

+++ web/src/org/openmrs/module/patientmatching/web/RecordSerializer.java
@@ -25,10 +25,12 @@ public class RecordSerializer {
         File configFileFolder = OpenmrsUtil.getDirectoryInApplicationDataDirectory(configLocation);
         File serialFile = new File(configFileFolder, filename);
         
-		String xml = stream.toXML(record);
-        BufferedWriter writer = new BufferedWriter(new FileWriter(serialFile));
-        writer.write(xml);
-        writer.close();
+        if (!serialFile.exists()) {
+    		String xml = stream.toXML(record);
+            BufferedWriter writer = new BufferedWriter(new FileWriter(serialFile));
+            writer.write(xml);
+            writer.close();
+        }
 	}
 	
 	public static Record deserialize(String xmlName) throws IOException {
